Jwt Interceptor used to hold and pass the user token (S9,V102)
When we use Jwt or anything token based authorization then every time we need to pass the Authorization token when calling an API from an angular client project. Then in everywhere where we call the API then need to pass that token. To reduce this redundancy interceptor used , it holds that Authorization in project when user successfully logged in.